refactor(layout.tsx, page.tsx): integrate ProfilePage into main page component for improved structure style(layout.tsx): add background color to navigation div for better UI consistency chore(page.tsx): remove redundant ProfilePage component file to reduce duplication
